+++ apps/codecs/mpc.c
@@ -116,7 +116,7 @@ next_track:
         goto exit;
     }
     frequency = info.sample_freq;
-    ci->configure(DSP_SET_FREQUENCY, (long *)info.sample_freq);
+    ci->configure(DSP_SET_FREQUENCY, (long *)(long)info.sample_freq);
         
     /* set playback engine up for correct number of channels */
     /* NOTE: current musepack format only allows for stereo files
